==> Rmd Files/2-prune_phylo.Rmd <==
---
title: "Prune amphibian phylogeny to salamander dataset"
author: "Nicole Bednarik and Katie Thomas"
date: 23 Jan 2025
output:
  html_document:
    code_fold: show
    theme: flatly
    toc: no
    toc_float: no
---

<style type="text/css">

body{ /* Normal  */
      font-size: 17px;
  }
  
</style>


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#load packages
library(ape)
library(geiger)
library(picante)
library(caper)
library(phytools)
library(knitr)
library(tidyverse)
library(dplyr)
```

# Load data

Here we load the list of species that we sampled eye size for - the subset that we want to trim the phylogeny to match. 

```{r load-data, results = "hide"}

# Load salamander data
caudata_raw <- read.csv("../Data/Tidy/salamanders_tidy.csv", header=TRUE, na.strings=c("", "NA", " ")) %>%
  # Remove the gilled version of Pleurodeles_waltl
  filter(Genus_Species != "Pleurodeles_waltl_gilled") %>%
  # Remove blind salamanders from dataset
  filter(Genus_Species != "Eurycea_rathbuni" & Genus_Species != "Proteus_anguinus")

# Load salamander data that includes blind salamanders (for Figure 1 for paper)
caudata_raw_blind <- read.csv("../Data/Tidy/salamanders_tidy.csv", header=TRUE, na.strings=c("", "NA", " ")) %>%
  # Remove the gilled version of Pleurodeles_waltl
  filter(Genus_Species != "Pleurodeles_waltl_gilled")

# Check out structure
str(caudata_raw)

# Tidy adult data
tip_species <- caudata_raw %>% 
  dplyr::select(Genus_Species, Family) %>%
  unique()
tip_species_blind <- caudata_raw_blind %>%
  dplyr::select(Genus_Species, Family) %>%
  unique()

# Check for duplicates of species
n_occur <- data.frame(table(tip_species$Genus_Species))
n_occur_blind <- data.frame(table(tip_species_blind$Genus_Species))
```

# Load published phylogeny

Here we import an amphibian tree published by Jetz and Pyron (2019). It's important to note that this tree includes branches supported by molecular data as well as branches that are grafted on based on taxonomy. 

We check whether the tree is rooted (it is) and whether it is reading as ultrametric (it should be, but is not, so we force it using force.ultrametric) and then look at the full tree. 

```{r pryontree, fig.height=8, fig.width=8, echo = TRUE}

# Import ultrametric phylogeny (modified) from Jetz and Pyron 2019
tree_orig <- read.tree(file = "../Data/Raw/amph_shl_new_Consensus_7238.tre") #reads tree

# Check whether tree is rooted
is.rooted(tree_orig) 

# Check whether tree is dichotomous (no polytomies)
is.binary(tree_orig) 

# Check that tree is ultrametric
is.ultrametric(tree_orig)

# Force ultrametric
tree <- force.ultrametric(tree_orig)

# Check that tree is ultrametric
is.ultrametric(tree)

# Plot tree
plot.phylo(tree, show.tip.label = F)
```

# Match names in tree and phylogeny

First, we find species that are in our dataset that aren't found as exact matches in the phylogeny tip labels. 

```{r match-names, results = "hide"}

# Find species in dataset that don't exist as phylogeny tip label
tip_species[which(!tip_species$Genus_Species %in% as.vector(tree$tip.label)), ]
tip_species_blind[which(!tip_species_blind$Genus_Species %in% as.vector(tree$tip.label)), ]
```

There is 1 species that doesn't appear in the phylogeny, _Aneides niger_. We can look for a synonym or close relative in the phylogeny tips. 

```{r phylo-tips, results = "hide"}

# Look for close relative in phylogeny tips
phylo.tips <- as.data.frame(tree$tip.label)
```

After looking at Frost (2022) and searching through the phylogeny tips, _Aenides niger_ is not found in the tree, but it previously was a subspecies of _Aneides flavipunctatus_, which is present in the tree. We also have that species in our dataset, so we will graft on a branch that is sister to A. flavipunctatus for A. niger. 

```{r graft-branch, results = "hide"}

# Copy tree for editing
tree_graft <- tree

# Find most recent common ancestor of Aneides flavipunctatus and sister clade of Aneides_vagrans + Aneides_ferreus
bind.node <- findMRCA(tree_graft, 
                      tips = c("Aneides_flavipunctatus", "Aneides_vagrans", "Aneides_ferreus"), 
                      type = "node")

# Add A. niger to that node
tree_graft2 <- bind.tip(tree = tree_graft, 
                        tip.label = "Aneides_niger", 
                        where = bind.node)
```

Now we can check that our names in the dataset are matching those in the phylogeny. Now that we've consolidated subspecies into species, we have 155 species with data that we want to match the tree (157 species for the second phylogeny that will include 2 blind salamander species).

```{r tree-match, results = "hide"}

# Find difference between tree and dataset
rownames(tip_species) <- tip_species$Genus_Species
overlap <- name.check(tree_graft2, tip_species)

rownames(tip_species_blind) <- tip_species_blind$Genus_Species
overlap_blind <- name.check(tree_graft2, tip_species_blind)

# Species in data that aren't matching tree
overlap$data_not_tree
overlap_blind$data_not_tree
```

All the species are now matching! Now we prune the tree to match the species in our dataset. 

```{r prune-tree, results = "hide"}

## Produce tree matching species with eye size data

# Make list of taxa to drop (in tree but not in data set)
drops <- setdiff(tree_graft2$tip.label, tip_species$Genus_Species)

# Drop unwanted tips from phylogeny
tree.pruned <- drop.tip(phy = tree_graft2, tip = drops) 

# See which tips you've kept in your phylogeny
tree.pruned$tip.label

# Check that phylogeny tips and data match exactly (if they match will return "OK")
name.check(tree.pruned, tip_species)

## Produce tree that includes eye size data species plus two blind species (no eyes)

# Make list of taxa to drop (in tree but not in data set)
drops_blind <- setdiff(tree_graft2$tip.label, tip_species_blind$Genus_Species)

# Drop unwanted tips from phylogeny
tree.pruned_blind <- drop.tip(phy = tree_graft2, tip = drops_blind) 

# See which tips you've kept in your phylogeny
tree.pruned_blind$tip.label
  
# Check that phylogeny tips and data match exactly (if they match will return "OK")
name.check(tree.pruned_blind, tip_species_blind)
```

The pruned tree tip labels and the dataset species names match exactly. We can plot the tree to check that it looks good. 

```{r, fig.height=25, fig.width=9}

# Plot pruned tree with polytomies
plot.phylo(tree.pruned, show.tip.label = TRUE)
plot.phylo(tree.pruned_blind, show.tip.label = TRUE)
```

We also check that the tree is still rooted and ultrametric. 

```{r check-tree, results = "hide"}

# Confirm that tree is rooted
is.rooted(tree.pruned) 
is.rooted(tree.pruned_blind)

# Check that tree is still ultrametric
is.ultrametric(tree.pruned)
is.ultrametric(tree.pruned_blind)
```

Finally, because we added a polytomy our tree will not be binary. For analyses, we will need it to be, so we can randomly resolve polytomies by adding in tiny differences so that we can use statistical methods like PGLS. 

```{r tree-bi, results = "hide"}

# Test whether tree is dichotomous (shouldn't be yet)
is.binary(tree.pruned)
is.binary(tree.pruned_blind)

# Randomly resolve polytomies to make tree dichotomous using multi2di function in ape
dich.tree <- multi2di(tree.pruned)
dich.tree_blind <- multi2di(tree.pruned_blind)

# Check that tree is now binary
is.binary(dich.tree)
is.binary(dich.tree_blind)
```

The tree is now dichotomous. We will take one final look at the tree to check that it looks ok and then export as a nexus file for use in our analyses. 

```{r, fig.height=25, fig.width=9}

# Plot final tree
plot.phylo(dich.tree, show.tip.label = TRUE)
# Plot final tree (with blind species)
plot.phylo(dich.tree_blind, show.tip.label = TRUE)
```

Looks good! Now we export. 

```{r, results = "hide"}

# Export final tree
write.nexus(dich.tree, file = "../Data/Tidy/caudata-tree.nex")
# Export final tree (with blind salamanders)
write.nexus(dich.tree_blind, file = "../Data/Tidy/caudata-tree-blind.nex")
```
