==> Rmd Files/3-allometry.Rmd <==
---
title: "Salamander eye scaling and investment"
author: "Nicole Bednarik & Katie Thomas"
date: 23 Jan 2025
output:
  html_document:
    theme: flatly
    toc: TRUE
    toc_float: TRUE
    code_fold: show
---


```{r setup, include = FALSE}

# Load package libraries
library(phytools)
library(dplyr)
library(smatr)
library(nlme)
library(ape)
library(geiger)
library(caper)
library(plotly) 
library(gridExtra)
library(mvMORPH)
library(tidyverse) 
library(knitr)

# Markdown settings
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, fig.height = 4.5, fig.width=8)

```

# Data

First we import our tidied salamander morphology & trait dataset. 

```{r load-data, results = "hide"}

# Import tidied salamander eye size and trait data
salamanders <- data.frame(read.csv("../Data/Tidy/salamanders_tidy.csv", header=TRUE, na.strings=c("", "NA", " ","<1")))

# Quick look at data structure
str(salamanders)
```

Next, we import the amphibian tree from Jetz and Pyron (2019) that has been pruned to match the salamander species in this dataset.

```{r, fig.height=18, fig.width=7}

# Import pruned phylogeny
caudatatree <- read.nexus("../Data/Tidy/caudata-tree.nex")

# Plot tree
plot.phylo(caudatatree, show.tip.label=TRUE, cex = 0.7)

```

# Data exploration

First, we can take a preliminary look at all of our data so that we can see how it looks and whether we notice any obvious outliers that we might want to go back and check out. Here we plot out our measurements from the right side vs. the left side of each specimen so that we can make sure that we have consistent measurements throughout the dataset. We use the package plotly to create interactive plots - hover over a point to see the name of the species measured. 


## Right eyes vs. left eyes

We expect that measurements of left and right eyes and corneas should be similar within specimens. We can plot out our data to test for this and make sure we don't see any glaring issues. We can also fit a regression, which we would expect to have a slope near 1 and an intercept near 0 if left and right eye measurements are similar. 

```{r right-left, fig.height=8, fig.width=8, echo = TRUE}

#### right eye diameter vs. left eye diameter ####

# Linear model 
RLeye.lm <- lm(ED_right_mm ~ ED_left_mm, data = salamanders)
summary(RLeye.lm)

# Plot
RLeye.plot <- ggplot(salamanders, aes(x = ED_left_mm, y = ED_right_mm, text = Genus_Species)) +
 geom_point(alpha = 0.9) +
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +
  geom_abline(slope = coef(RLeye.lm)[[2]], intercept = coef(RLeye.lm)[[1]]) 


# Interactive plot
ggplotly(RLeye.plot)

```

As we would expect, measurements of the left and right eye are highly correlated, with a slope close to 1 and an intercept close to 0. We can do the same test for the corneal measurements. 

```{r left-right, fig.height=8, fig.width=8, echo = TRUE}

#### right cornea diameter vs. left cornea diameter ####

# Linear model 
RLcornea.lm <- lm(CD_right_mm ~ CD_left_mm, data = salamanders)
summary(RLcornea.lm)

# Plot
RLcor.plot <- ggplot(salamanders, aes(x = CD_left_mm, y = CD_right_mm, text = Genus_Species)) +
 geom_point(alpha = 0.9) +
  theme_bw() +
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +
  geom_abline(slope = coef(RLcornea.lm)[[2]], intercept = coef(RLcornea.lm)[[1]]) 


# Interactive plot
ggplotly(RLcor.plot)

```

Again, we find that left and right corneal measurements are highly correlated, with a slope close to 1 and an intercept close to 0. Together, these results indicate that measurements were consistent between the left and right eyes of a specimen, which is good! For looking at interspecific allometry, we will calculate species means from these data so that we can use phylogenetic comparative methods. 

# Interspecific allometry

Here, we will use phylogenetic generalised least-squares (PGLS) regressions to see how eyes are scaling with body size across different species of salamanders with different ecologies. PGLS analyses are similar to normal linear regressions, but also account for the non-independence of data collected from close relatives due to shared evolutionary history. 

In these analyses, we will log eye size and body size variables, as this is standard for allometric relationships. 
First, we will calculate the mean measurements across specimens for each species. We need to make sure in doing this that we omit rows with missing data for a variable of interest. In looking at the data, we can see that we have many more measurements for cornea diameter than for eye diameter.

## Species means for morphological comparisons

We then find species means for various morphological parameters. There are some specimens that lack data for eye size but have data for svl, mass, and corneal diameter (these measurements are present for all specimens). However, when eye size measurements are missing, they are consistently missing for all specimens measured within a species. This means we don't have to worry about missing data for one specimen corrupting the mean across several specimens for eye size (it is either present for all specimens in a species or absent for all specimens in a species). So we can make just one dataframe of species averages, and some species will just lack eye size data. 

```{r species-av, results = "hide"}

# Species means for eye size, cornea size, SVL, and rootmass data
av.sal <- salamanders %>% 
  mutate_if(is.character, as.factor) %>%
  group_by(Genus_Species) %>%
  summarise(eye_av = mean(eyemean), 
            cor_av = mean(cormean),
            svl_av = mean(SVL_mm), 
            rootmass_av = mean(rootmass), 
            n = n())

# Merge data means with other column info, keeping only trait columns
av.sal <- merge(av.sal, salamanders[match(unique(salamanders$Genus_Species), salamanders$Genus_Species), ], by="Genus_Species", all.x = TRUE, all.y = FALSE) %>% 
  select(Genus_Species, eye_av, cor_av, svl_av, rootmass_av, n, Order, Suborder, Family, Genus, Species, Gill_Presence, Activity_Period, Adult_Habitat, Development, Life_History, Larval_Habitat) 

# Remove Eurycea rathbuni and Proteus anguinus (blind salamanders) from dataframe
av.sal <- av.sal[-c(77, 133),]
# Export tidied dataset for analyses
write.csv(av.sal, file = "../Data/Tidy/salamanders_averages.csv")
```

# Color figures with pallettes in Thomas et al. 2020

Here we run a custom color palette for the adult habitat categories to be used for all following figures. 

First we define what we want that color palette to be:

```{r color-pallette}

# See what the levels of adult habitat are named
levels(as.factor(av.sal$Adult_Habitat))

# Define a colorblind-friendly vector of colors for adult habitat
col_hab <- c("aquatic" = "#0072B2",
             "scansorial" = "#009E73",
             "semiaquatic" = "#56B4E9",
             "subfossorial" = "#CC79A7",
             "terrestrial" = "#E69F00")

# Now you can see that each habitat is assigned a hex color when you look at the vector
col_hab

# Make a vector of point shapes for adult habitat
shape_hab <- c("aquatic" = 25,
             "scansorial" = 24,
             "semiaquatic" = 15,
             "subfossorial" = 16,
             "terrestrial" = 18)
#see that each state has a shape now
shape_hab
```

## Phylogenetic generalised least squares regressions

In order to fit a PGLS regression in caper, we first need to make a comparative data object that includes our dataset and our phylogeny. Note that the gilled Pleurodeles waltl will be dropped from this, as there can only be one row for each species. 

```{r merge-data, results = "hide"}

# Make row names of dataset the species names (so it will match phylogeny tips)
rownames(av.sal) <- av.sal$Genus_Species

# Check that names match in dataframe and tree
name.check(phy = caudatatree, data = av.sal)

# Use caper function to combine phylogeny and data into one object (this function also matches species names in tree and dataset)
sal.comp <- comparative.data(phy = caudatatree, data = av.sal, 
                            names.col = Genus_Species, vcv = TRUE, 
                            na.omit = FALSE, warn.dropped = TRUE)

# Check for dropped tips or dropped species
sal.comp$dropped$tips #phylogeny
sal.comp$dropped$unmatched.rows #dataset

```

Here we fit PGLS regressions and plot the resulting outputs for the logged eye size and body size variables.

### PGLS of log eye diameter vs. log cube root of mass

```{r pgls-edmass}

# PGLS eye diameter vs. cube root of mass
pgls_edmass <- pgls(log10(eye_av) ~ log10(rootmass_av), 
               data = sal.comp, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

# Diagnostic plots
par(mar = c(4,4,2,2))
par(mfrow = c(2, 2))
plot(pgls_edmass)
par(mfrow = c(1, 1))

# Likelihood plot for Pagel's lambda. Solid red line indicates estimate for lambda and broken red lines indicate the 95% confidence interval
lambda.edmass <- pgls.profile(pgls_edmass, "lambda")
plot(lambda.edmass)

# Print model output 
summary(pgls_edmass)

# Quick plot of pgls
plot(log10(eye_av) ~ log10(rootmass_av), data = av.sal)
abline(pgls_edmass)
```

### Plot of eye diameter versus cube root of mass

```{r plot-edmass} 

# Plot average eye diameter (mm) against cube root of mass (g)
## Size and dimensions modified for manuscript  
plot_pgls_edmass <- ggplot(av.sal, aes(y=eye_av, x=rootmass_av, text=Genus_Species)) +
  geom_point(alpha = 0.9, aes(color = Adult_Habitat), size = 5, na.rm = TRUE) + 
  scale_color_manual(values = col_hab, name = "Adult habitat") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    axis.text.x = element_text(size = 26),  
    axis.text.y = element_text(size = 26), 
    axis.title.x = element_text(size = 30, color = "black", margin = margin(t = 10)), 
    axis.title.y = element_text(size = 30, color = "black", margin = margin(r = 18)), 
    axis.ticks.length = unit(0.3, "cm"), 
    panel.border = element_rect(size = 2)  
  ) +
  scale_y_log10(name = "Eye diameter (mm)", breaks = c(2, 3, 7)) +
  scale_x_log10(name = "Cube root of mass (g)", breaks = c(1, 3, 5), limits = c(0.5, 5.5)) +
  geom_abline(slope = coef(pgls_edmass)[[2]], intercept = coef(pgls_edmass)[[1]], linetype = "solid", size = 1.5) 

ggplotly(plot_pgls_edmass)

# Save the plot as a rectangle with specified width and height
ggsave("../Plots/plot_pgls_edmass.png", plot = plot_pgls_edmass, width = 9, height = 7.5, units = "in")

```


### PGLS of log eye diameter vs. log snout-vent length 

```{r pgls-edsvl}

# PGLS eye diameter vs. SVL
pgls_edsvl <- pgls(log10(eye_av) ~ log10(svl_av), 
               data = sal.comp, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

# Diagnostic plots
par(mar = c(4,4,2,2))
par(mfrow = c(2, 2))
plot(pgls_edsvl)
par(mfrow = c(1, 1))

# Likelihood plot for Pagel's lambda. Solid red line indicates estimate for lambda and broken red lines indicate the 95% confidence interval
lambda.edsvl <- pgls.profile(pgls_edsvl, "lambda")
plot(lambda.edsvl)

# Print model output 
summary(pgls_edsvl)

# Quick plot of pgls
plot(log10(eye_av) ~ log10(svl_av), data = av.sal)
abline(pgls_edsvl)
```

### Plot of eye diameter versus snout-vent length

```{r plot-edsvl} 

# Plot average eye diameter (mm) against snout-vent length (mm)
## Size and dimensions modified for manuscript  
plot_pgls_edsvl <- ggplot(av.sal, aes(y=eye_av, x=svl_av, text=Genus_Species)) +
  geom_point(alpha = 0.9, aes(color = Adult_Habitat), size = 5, na.rm = TRUE) +  
  scale_color_manual(values = col_hab, name = "Adult habitat") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    axis.text.x = element_text(size = 26),  
    axis.text.y = element_text(size = 26),  
    axis.title.x = element_text(size = 30, color = "black", margin = margin(t = 10)),  
    axis.title.y = element_text(size = 30, color = "black", margin = margin(r = 18)),  
    axis.ticks.length = unit(0.3, "cm"), 
    panel.border = element_rect(size = 2) 
  ) +
  scale_y_log10(name = "Eye diameter (mm)", breaks = c(2, 3, 7)) +
  scale_x_log10(name = "Snout-vent length (mm)", breaks = c(30, 70, 150), limits = c(25,160)) +
  geom_abline(slope = coef(pgls_edsvl)[[2]], intercept = coef(pgls_edsvl)[[1]], linetype = "solid", size = 1.5)

ggplotly(plot_pgls_edsvl)

# Save the plot as a rectangle with specified width and height
ggsave("../Plots/plot_pgls_edsvl.png", plot = plot_pgls_edsvl, width = 9, height = 7.5, units = "in")
```

### PGLS of log eye diameter vs. log cornea diameter

```{r pgls-edcd}

# PGLS eye diameter vs. cornea diameter
pgls_edcd <- pgls(log10(cor_av) ~ log10(eye_av), 
               data = sal.comp, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

# Diagnostic plots
par(mar = c(4,4,2,2))
par(mfrow = c(2, 2))
plot(pgls_edcd)
par(mfrow = c(1, 1))

# Likelihood plot for Pagel's lambda. Solid red line indicates estimate for lambda and broken red lines indicate the 95% confidence interval
lambda.edcd <- pgls.profile(pgls_edcd, "lambda")
plot(lambda.edcd)

# Print model output 
summary(pgls_edcd)

# Quick plot of pgls
plot(log10(cor_av) ~ log10(eye_av), data = av.sal)
abline(pgls_edcd)
```

### Plot of eye diameter versus cornea diameter

```{r plot-edcd} 

# Plot average eye diameter (mm) against cornea diameter (mm)
## Size and dimensions modified for manuscript  
plot_pgls_edcd <- ggplot(av.sal, aes(y=cor_av, x=eye_av, text=Genus_Species)) +
  geom_point(alpha = 0.9, aes(color = Adult_Habitat), size = 5, na.rm = TRUE) +  
  scale_color_manual(values = col_hab, name = "Adult habitat") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    axis.text.x = element_text(size = 26), 
    axis.text.y = element_text(size = 26), 
    axis.title.x = element_text(size = 30, color = "black", margin = margin(t = 10)), 
    axis.title.y = element_text(size = 30, color = "black", margin = margin(r = 18)), 
    axis.ticks.length = unit(0.3, "cm"),  
    panel.border = element_rect(size = 2) 
  ) +
  scale_y_log10(name = "Cornea diameter (mm)") +
  scale_x_log10(name = "Eye diameter (mm)", breaks = c(2, 3, 7)) +
  geom_abline(slope = coef(pgls_edcd)[[2]], intercept = coef(pgls_edcd)[[1]], linetype = "solid", size = 1.5)

ggplotly(plot_pgls_edcd)

# Save the plot as a rectangle with specified width and height
ggsave("../Plots/plot_pgls_edcd.png", plot = plot_pgls_edcd, width = 9, height = 7.5, units = "in")
```

### PGLS of log cornea diameter vs. log cube root of mass

```{r pgls-cdmass}

# PGLS cornea diameter vs. cube root of mass
pgls_cdmass <- pgls(log10(cor_av) ~ log10(rootmass_av),
               data = sal.comp, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

# Diagnostic plots
par(mar = c(4,4,2,2))
par(mfrow = c(2, 2))
plot(pgls_cdmass)
par(mfrow = c(1, 1))

# Likelihood plot for Pagel's lambda. Solid red line indicates estimate for lambda and broken red lines indcaite the 95% confidence interval
lambda.cdmass <- pgls.profile(pgls_cdmass, "lambda")
plot(lambda.cdmass)

# Print model output 
summary(pgls_cdmass)

# Quick plot of pgls
plot(log10(cor_av) ~ log10(rootmass_av), data = av.sal)
abline(pgls_cdmass)
```

### Plot of cornea diameter versus cube root of mass

```{r plot-cdmass} 

# Plot average cornea diameter (mm) against cube root of mass (g)
## Size and dimensions modified for manuscript  
plot_pgls_cdmass <- ggplot(av.sal, aes(y=cor_av, x=rootmass_av, text=Genus_Species)) +
  geom_point(alpha = 0.9, aes(color = Adult_Habitat), size = 5, na.rm = TRUE) +  
  scale_color_manual(values = col_hab, name = "Adult habitat") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    axis.text.x = element_text(size = 26),  
    axis.text.y = element_text(size = 26), 
    axis.title.x = element_text(size = 30, color = "black", margin = margin(t = 10)), 
    axis.title.y = element_text(size = 30, color = "black", margin = margin(r = 18)), 
    axis.ticks.length = unit(0.3, "cm"),  
    panel.border = element_rect(size = 2) 
  ) +
  scale_y_log10(name = "Cornea diameter (mm)") +
  scale_x_log10(name = "Cube root of mass (g)", breaks = c(1, 3, 5), limits = c(0.5, 5.5)) +
  geom_abline(slope = coef(pgls_cdmass)[[2]], intercept = coef(pgls_cdmass)[[1]], linetype = "solid", size = 1.5)

ggplotly(plot_pgls_cdmass)

# Save the plot as a rectangle with specified width and height
ggsave("../Plots/plot_pgls_cdmass.png", plot = plot_pgls_cdmass, width = 9, height = 7.5, units = "in")
```

### PGLS of log cornea diameter vs. log snout-vent length 

```{r pgls-cdsvl}

# PGLS cornea diameter vs. SVL
pgls_cdsvl <- pgls(log10(cor_av) ~ log10(svl_av),
               data = sal.comp, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

# Diagnostic plots
par(mar = c(4,4,2,2))
par(mfrow = c(2, 2))
plot(pgls_cdsvl)
par(mfrow = c(1, 1))

# Likelihood plot for Pagel's lambda. Solid red line indicates estimate for lambda and broken red lines indcaite the 95% confidence interval
lambda.cdsvl <- pgls.profile(pgls_cdsvl, "lambda")
plot(lambda.cdsvl)

# Print model output 
summary(pgls_cdsvl)

# Quick plot of pgls
plot(log10(cor_av) ~ log10(svl_av), data = av.sal)
abline(pgls_cdsvl)
```

### Plot of cornea diameter versus snout-vent length

```{r plot-cdsvl} 

# Plot average cornea diameter (mm) against snout-vent length (mm)
## Size and dimensions modified for manuscript  
plot_pgls_cdsvl <- ggplot(av.sal, aes(y=cor_av, x=svl_av, text=Genus_Species)) +
  geom_point(alpha = 0.9, aes(color = Adult_Habitat), size = 5, na.rm = TRUE) +  
  scale_color_manual(values = col_hab, name = "Adult habitat") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    axis.text.x = element_text(size = 26), 
    axis.text.y = element_text(size = 26), 
    axis.title.x = element_text(size = 30, color = "black", margin = margin(t = 10)),  
    axis.title.y = element_text(size = 30, color = "black", margin = margin(r = 18)),  
    axis.ticks.length = unit(0.3, "cm"),  
    panel.border = element_rect(size = 2) 
  ) +
  scale_y_log10(name = "Cornea diameter (mm)") +
  scale_x_log10(name = "Snout-vent length (mm)", breaks = c(30, 70, 150), limits = c(25,160)) +
  geom_abline(slope = coef(pgls_cdsvl)[[2]], intercept = coef(pgls_cdsvl)[[1]], linetype = "solid", size = 1.5)

ggplotly(plot_pgls_cdsvl)

# Save the plot as a rectangle with specified width and height
ggsave("../Plots/plot_pgls_cdsvl.png", plot = plot_pgls_cdsvl, width = 9, height = 7.5, units = "in")
```

### PGLS of log snout-vent length vs. log cube root of mass

```{r pgls-svlmass}

# PGLS svl vs. cube root of mass
pgls_svlmass <- pgls(log10(svl_av) ~ log10(rootmass_av),
               data = sal.comp, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

# Diagnostic plots
par(mar = c(4,4,2,2))
par(mfrow = c(2, 2))
plot(pgls_svlmass)
par(mfrow = c(1, 1))

# Likelihood plot for Pagel's lambda. Solid red line indicates estimate for lambda and broken red lines indcaite the 95% confidence interval
lambda.svlmass <- pgls.profile(pgls_svlmass, "lambda")
plot(lambda.svlmass)

# Print model output 
summary(pgls_svlmass)

# Quick plot of pgls
plot(log10(svl_av) ~ log10(rootmass_av), data = av.sal)
abline(pgls_svlmass)
```

### Plot of snout-vent length versus cube root of mass

```{r plot-svlmass} 

# Plot average snout-vent length (mm) against cube root of mass (g)
## Size and dimensions modified for manuscript  
plot_pgls_svlmass <- ggplot(av.sal, aes(y=svl_av, x=rootmass_av, text=Genus_Species)) +
  geom_point(alpha = 0.9, aes(color = Adult_Habitat), size = 5, na.rm = TRUE) +  
  scale_color_manual(values = col_hab, name = "Adult habitat") +
  theme_bw() +
  theme(
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    axis.text.x = element_text(size = 26.3), 
    axis.text.y = element_text(size = 26.3), 
    axis.title.x = element_text(size = 30.3, color = "black", margin = margin(t = 10)),  
    axis.title.y = element_text(size = 30.3, color = "black", margin = margin(r = 12)),  
    axis.ticks.length = unit(0.3, "cm"),  
    panel.border = element_rect(size = 2)  
  ) +
  scale_y_log10(name = "Snout-vent length (mm)", breaks = c(30, 70, 150), limits = c(25,160)) +
  scale_x_log10(name = "Cube root of mass (g)", breaks = c(1, 3, 5), limits = c(0.5,5.5)) +
  geom_abline(slope = coef(pgls_svlmass)[[2]], intercept = coef(pgls_svlmass)[[1]], linetype = "solid", size = 1.5)

ggplotly(plot_pgls_svlmass)

# Save the plot as a rectangle with specified width and height
ggsave("../Plots/plot_pgls_svlmass.png", plot = plot_pgls_svlmass, width = 9.4, height = 7.5, units = "in")
```
