==> Rmd Files/4-eye_investment.Rmd <==
---
title: "Salamander eye scaling and investment"
author: "Nicole Bednarik & Katie Thomas"
date: 20 Sep 2024
output:
  html_document:
    theme: flatly
    toc: TRUE
    toc_float: TRUE
    code_fold: show
---

```{r setup, include = FALSE}

# Load package libraries
library(plyr)
library(smatr)
library(picante)
library(nlme)
library(ape)
library(caper)
library(phytools)
library(plotly)
library(gridExtra)
library(geiger)
library(ggtree)
library(cowplot)
library(knitr)
library(mvMORPH)
library(tidyverse)

# Markdown settings
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, fig.height = 4.5, fig.width=8)

```

# Data

First we import our tidied salamander morphology & trait dataset. 

```{r load-data, results = "hide"}

# Import tidied salamander eye size and trait data
av.sal <- data.frame(read.csv("../Data/Tidy/salamanders_averages.csv", header=TRUE, na.strings=c("", "NA", " ","<1")))

# Quick look at data structure
str(av.sal)

```

Next, we import the amphibian tree from Jetz and Pyron (2019) that has been pruned to match the salamander species in this dataset.

```{r, fig.height=18, fig.width=7}

# Import pruned phylogeny
caudatatree <- read.nexus("../Data/Tidy/caudata-tree.nex")

# Plot tree
plot.phylo(caudatatree, show.tip.label=TRUE, cex = 0.7)

```

# Interspecific allometry using PGLS

In order to fit a PGLS regression in caper, we first need to make a comparative data object that includes our dataset and our phylogeny. Note that the gilled Pleurodeles waltl will be dropped from this, as there can only be one row for each species. 

```{r merge-dataset, results = "hide"}

# Make row names of dataset the species names (so it will match phylogeny tips)
rownames(av.sal) <- av.sal$Genus_Species

# Check that names match in dataframe and tree
name.check(phy = caudatatree, data = av.sal)

# Use caper function to combine phylogeny and data into one object (this function also matches species names in tree and dataset)
sal.comp <- comparative.data(phy = caudatatree, data = av.sal, 
                            names.col = Genus_Species, vcv = TRUE, 
                            na.omit = FALSE, warn.dropped = TRUE)

# Check for dropped tips or dropped species
sal.comp$dropped$Genus_Species#phylogeny
sal.comp$dropped$unmatched.rows #dataset

```

# Relative eye size and eye investment in salamanders

To investigate how relative eye size differs among species with different ecological traits, we look at the residuals from PGLS fits of eye size versus body size variables. 

First, we add residuals of PGLS fits to our dataset by species.

```{r pgls-residuals, results = "hide"}

# PGLS cornea diameter vs. cube root of mass ------------
pgls_cdmass <- pgls(log10(cor_av) ~ log10(rootmass_av), 
               data = sal.comp, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

# Residuals for PGLS of cornea diameter ~ cube root of mass

# Extract PGLS residuals 
pglsres.cdmass <- residuals(pgls_cdmass) 

# Name residuals
colnames(pglsres.cdmass) <- "pglsres.cdmass" 

# Making PGLS vector into a dataframe
pglsres.cdmass <- as.data.frame(pglsres.cdmass)

# Make rownames into a column so we can use merge command
pglsres.cdmass <- tibble::rownames_to_column(pglsres.cdmass, "rn")

# Rename Genus_Species by making a new column then deleting the old one
pglsres.cdmass$Genus_Species <- pglsres.cdmass$rn
pglsres.cdmass$rn <- NULL

# Merge datasets
pglsres.cdmass.all <- merge(av.sal, pglsres.cdmass, by = "Genus_Species")

# Export tidied dataset for analyses
write.csv(pglsres.cdmass.all, file = "../Data/Tidy/pglsres.cdmass.all.csv")

```

```{r, results = "hide"}

#PGLS eye diameter vs. cube root of mass ------------
pgls_edmass <- pgls(log10(eye_av) ~ log10(rootmass_av), 
               data = sal.comp, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

#Residuals for PGLS of eye diameter ~ cube root of mass

# Extract PGLS residuals 
pglsres.edmass <- residuals(pgls_edmass) 

# Name residuals
colnames(pglsres.edmass) <- "pglsres.edmass" 

# Making PGLS vector into a dataframe
pglsres.edmass <- as.data.frame(pglsres.edmass)

# Make rownames into a column so we can use merge command
pglsres.edmass <- tibble::rownames_to_column(pglsres.edmass, "rn")

# Renaming Genus_Species by making a new column then deleting the old one
pglsres.edmass$Genus_Species <- pglsres.edmass$rn
pglsres.edmass$rn <- NULL

# Merging datasets
pglsres.edmass.all <- merge(av.sal, pglsres.edmass, by = "Genus_Species")

# Export tidied dataset for analyses
write.csv(pglsres.edmass.all, file = "../Data/Tidy/pglsres.edmass.all.csv")
```

```{r, results = "hide"}

# PGLS cornea diameter vs. SVL ------------
pgls_cdsvl <- pgls(log10(cor_av) ~ log10(svl_av), 
               data = sal.comp, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

#Residuals for PGLS of cornea diameter ~ snout-vent length

# Extract PGLS residuals 
pglsres.cdsvl <- residuals(pgls_cdsvl) 

# Name residuals
colnames(pglsres.cdsvl) <- "pglsres.cdsvl" 

# Making PGLS vector into a dataframe
pglsres.cdsvl <- as.data.frame(pglsres.cdsvl)

# Make rownames into a column so we can use merge command
pglsres.cdsvl <- tibble::rownames_to_column(pglsres.cdsvl, "rn")

# Renaming Genus_Species by making a new column then deleting the old one
pglsres.cdsvl$Genus_Species <- pglsres.cdsvl$rn
pglsres.cdsvl$rn <- NULL

# Merging datasets
pglsres.cdsvl.all <- merge(av.sal, pglsres.cdsvl, by = "Genus_Species")

# Export tidied dataset for analyses
write.csv(pglsres.cdsvl.all, file = "../Data/Tidy/pglsres.cdsvl.all.csv")
```

```{r, results = "hide"}

# PGLS eye diameter vs. SVL ------------
pgls_edsvl <- pgls(log10(eye_av) ~ log10(svl_av), 
               data = sal.comp, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

# Residuals for PGLS of eye diameter ~ snout-vent length

# Extract pgls residuals 
pglsres.edsvl <- residuals(pgls_edsvl) 

# Name residuals
colnames(pglsres.edsvl) <- "pglsres.edsvl" 

# Making PGLS vector into a dataframe
pglsres.edsvl <- as.data.frame(pglsres.edsvl)

# Make rownames into a column so we can use merge command
pglsres.edsvl <- tibble::rownames_to_column(pglsres.edsvl, "rn")

# Renaming Genus_Species by making a new column then deleting the old one
pglsres.edsvl$Genus_Species <- pglsres.edsvl$rn
pglsres.edsvl$rn <- NULL

# Merging datasets
pglsres.edsvl.all <- merge(av.sal, pglsres.edsvl, by = "Genus_Species")

# Export tidied dataset for analyses
write.csv(pglsres.edsvl.all, file = "../Data/Tidy/pglsres.edsvl.all.csv")
```
