==> Rmd Files/7-PGLS.analysis.Rmd <==
---
title: "PGLS Analysis"
author: "Nicole Bednarik & Katie Thomas"
date: 25 October 2022
output:
  html_document:
    theme: flatly
    toc: TRUE
    toc_float: TRUE
    code_fold: show
---

```{r setup, include = FALSE}

# Load package libraries
library(plyr)
library(geiger)
library(caper)
library(plotly)
library(kableExtra)
library(gridExtra)
library(ggimage)
library(grid)
library(gtable)
library(cowplot)
library(knitr)
library(tidyverse)

# Markdown settings
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, fig.height = 4.5, fig.width=8)

```

# Data

First we import our tidied salamander morphology & trait binary dataset with the residuals of a cornea ~ cube root of mass PGLS as a column.

```{r load-data}

# Import tidied salamander eye size and trait data
sal.res <- data.frame(read.csv("../Data/Tidy/salamanders_residuals_binary3.csv", header=TRUE, na.strings=c("", "NA", " ","<1")))
sal.res.2 <- data.frame(read.csv("../Data/Tidy/salamanders_residuals_updated.csv", header=TRUE, na.strings=c("", "NA", " ", "<1")))
aquaticspecies <- data.frame(read.csv("../Data/Raw/aquaticspecies.csv", header=TRUE, na.strings=c("", "NA", " ", "<1")))
caudatatree <- read.nexus("../Data/caudata-tree")

#change activity period for sal.res2 to a binary nocturnal vs. not nocturnal
sal.res.new <- sal.res.2 %>%
  #make new column called activity2 and change "both" and "diurnal" to "non-nocturnal"
  mutate(activity2 = recode(Activity.Period, "both" = "non-nocturnal", "diurnal" = "non-nocturnal", "nocturnal" = "nocturnal"))


# Quick look at data structure
str(sal.res)
str(sal.res.2)



#Make row names of dataset the species names (so it will match phylogeny tips)
rownames(sal.res.2) <- sal.res.2$Genus_Species
#rownames(aquaticspecies) <-aquaticspecies$Genus_Species
rownames(sal.res.new) <- sal.res.new$Genus_Species
rownames(sal.res) <- sal.res$Genus_Species


#check that names match in dataframe and tree
name.check(phy = caudatatree, data = sal.res.2)
name.check(phy = caudatatree, data = sal.res.new)
name.check(phy = caudatatree, data = sal.res)
#name.check(phy = caudatatree, data = aquaticspecies)

drops <- name.check(phy = caudatatree, data = sal.res.2)
drops1 <- name.check(phy = caudatatree, data = sal.res.new)
drops2 <- name.check(phy = caudatatree, data = sal.res)
#drops2 <- name.check(phy = caudatatree, data = aquaticspecies)

#drop the blind salamander from the phylogeny for plots
tree.hab <- drop.tip(caudatatree, drops$tree_not_data)
tree.act <- drop.tip(caudatatree, drops$tree_not_data)
tree.pupil <- drop.tip(caudatatree, drops$tree_not_data)
#tree.aq <- drop.tip(caudatatree, drops$tree_not_data)

#use caper function to combine phylogeny and data into one object (this function also matches species names in tree and dataset)

#sal.res.2 <- factor(Larval.Habitat, levels=c("no larva", "lentic water", "lotic water", "both"))
sal.comp <- comparative.data(phy = tree.hab, data = sal.res.2, 
                            names.col = Genus_Species, vcv = TRUE, 
                            na.omit = FALSE, warn.dropped = TRUE)
sal.comp.1 <- comparative.data(phy = tree.act, data = sal.res.new, 
                            names.col = Genus_Species, vcv = TRUE, 
                            na.omit = FALSE, warn.dropped = TRUE)
sal.comp.2 <- comparative.data(phy = tree.pupil, data = sal.res, 
                            names.col = Genus_Species, vcv = TRUE, 
                            na.omit = FALSE, warn.dropped = TRUE)
#sal.comp.2 <- comparative.data(phy = tree.aq, data = aquaticspecies,
                            #names.col = Genus_Species, vcv = TRUE,
                            #na.omit = FALSE, warn.dropped = TRUE)

levels(factor(sal.res.2$Larval.Habitat))
sal.res.2 <- ordered(sal.res.2$Larval.Habitat, levels=c("no larva", "lentic water", "lotic water", "both"))

#check for dropped tips or dropped species
sal.comp$dropped$tips #phylogeny
sal.comp.1$dropped$tips #phylogeny
#sal.comp.2$dropped$tips

```

#Habitat PGLS

```{r}
# Run habitat pgls--------

pgls_hab1 <- pgls(cor_av ~ Adult.Habitat,
               data = sal.comp.1, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

#main effects
anova(pgls_hab1)

#print model output 
summary(pgls_hab1)

#PGLS cornea diameter vs. cube root of mass + habitat
pgls_hab <- pgls(log10(cor_av) ~ log10(rootmass_av) + Adult.Habitat,
               data = sal.comp.1, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

#diagnostic plots
par(mar = c(4,4,2,2))
par(mfrow = c(2, 2))
plot(pgls_hab)

par(mfrow = c(1, 1))

#Likelihood plot for Pagel's lambda. Solid red line indicates estimate for lambda and broken red lines indcaite the 95% confidence interval
lambda.hab <- pgls.profile(pgls_hab, "lambda")
plot(lambda.hab)
```

```{r}
# Run habitat pgls--------

pgls_hab1 <- pgls(cor_av ~ Adult.Habitat,
               data = sal.comp.1, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

#main effects
anova(pgls_hab1)

#print model output 
summary(pgls_hab1)

#PGLS cornea diameter vs. cube root of mass + habitat
pgls_hab <- pgls(log10(cor_av) ~ log10(rootmass_av) + Adult.Habitat,
               data = sal.comp.1, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

#diagnostic plots
par(mar = c(4,4,2,2))
par(mfrow = c(2, 2))
plot(pgls_hab)

par(mfrow = c(1, 1))

#Likelihood plot for Pagel's lambda. Solid red line indicates estimate for lambda and broken red lines indcaite the 95% confidence interval
lambda.hab <- pgls.profile(pgls_hab, "lambda")
plot(lambda.hab)
```

```{r}
#main effects
anova(pgls_hab)

#print model output 
summary(pgls_hab)
```
# Development PGLS

```{r}
# Run Development pgls--------

#PGLS cornea diameter vs. cube root of mass + development
pgls_dev <- pgls(log10(cor_av) ~ log10(rootmass_av) + Development,
               data = sal.comp.1, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

#diagnostic plots
par(mar = c(4,4,2,2))
par(mfrow = c(2, 2))
plot(pgls_dev)

par(mfrow = c(1, 1))

#Likelihood plot for Pagel's lambda. Solid red line indicates estimate for lambda and broken red lines indcaite the 95% confidence interval
lambda.dev <- pgls.profile(pgls_dev, "lambda")
plot(lambda.dev)

```

```{r}
#main effects
anova(pgls_dev)

#print model output 
summary(pgls_dev)
```

#Aquatic PGLS
```{r}
# Run Aquatic pgls--------

#PGLS cornea diameter vs. cube root of mass + adult habitat for aquatic species
pgls_aq <- pgls(log10(cor_av) ~ log10(rootmass_av) + Adult.Habitat,
               data = aquaticspecies, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

#diagnostic plots
par(mar = c(4,4,2,2))
par(mfrow = c(2, 2))
plot(pgls_aq)

par(mfrow = c(1, 1))

#Likelihood plot for Pagel's lambda. Solid red line indicates estimate for lambda and broken red lines indcaite the 95% confidence interval
lambda.aq <- pgls.profile(pgls_aq, "lambda")
plot(lambda.aq)

```

```{r}
#main effects
anova(pgls_aq)

#print model output 
summary(pgls_aq)
```

#Activity Period

##Activity Period Plot

```{r}
#see what the levels of activity period are named
levels(as.factor(sal.res.new$activity2))

#define a colorblind-friendly vector of colors for adult habitat
col_act <- c("nocturnal" = "#330099",
             "non-nocturnal" = "#9966FF")

#now you can see that each level is assigned a hex color when you look at the vector
col_act

#make a vector of point shapes for levels
shape_act <- c("nocturnal" = 25,
             "non-nocturnal" = 24)

#see that each state has a shape now
shape_act
```

```{r}
#boxplot of pgls residuals vs. activity period (relative eye size)
activityperiod.rel <- ggplot(data = filter(sal.res.new, !is.na(Activity.Period)), aes(x = reorder(activity2, pglsres.cdmass, fun=mean), y=pglsres.cdmass, fill=activity2, shape=activity2)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA, alpha=0.5) +
  stat_summary(fun.y = mean, colour="black", geom="point", size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species, shape = activity2), size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Relative cornea size") +
  xlab("Activity Period") +
  theme(legend.position = "none") +
# use our custom colors for activity period
  scale_fill_manual(values = col_act, name = "Activity Period") +
# use our custom shapes for points
  scale_shape_manual(values = shape_act, name = "Activity Period") 

#print plot
activityperiod.rel

#look at the plot with the plotly package - you can hover over data points to see what each species is, great way to check outliers!
library(plotly)
ggplotly(activityperiod.rel)

```
##Activity Period PGLS

```{r}
# Run Activity Period pgls--------

#PGLS cornea diameter vs. cube root of mass + activity period
pgls_act <- pgls(log10(cor_av) ~ log10(rootmass_av) + activity2,
               data = sal.comp.1, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

#diagnostic plots
par(mar = c(4,4,2,2))
par(mfrow = c(2, 2))
plot(pgls_act)

par(mfrow = c(1, 1))

#Likelihood plot for Pagel's lambda. Solid red line indicates estimate for lambda and broken red lines indcaite the 95% confidence interval
lambda.act <- pgls.profile(pgls_act, "lambda")
plot(lambda.act)

```

```{r}
#main effects
anova(pgls_act)

#print model output 
summary(pgls_act)
```

#Life History

##Life History Plot

```{r}
#see what the levels of development are named
levels(as.factor(sal.res.2$Life.History))

#define a colorblind-friendly vector of colors for adult habitat
col_lh <- c("free-living" = "#006600",
             "no free-living" = "#66FF33")

#now you can see that each habitat is assigned a hex color when you look at the vector
col_lh

#make a vector of point shapes for adult habitat
shape_lh <- c("free-living" = 20,
             "no free-living" = 21)

#see that each state has a shape now
shape_lh
```

```{r}
#boxplot of pgls residuals vs. development (relative eye size)
lifehistory.rel <- ggplot(data = filter(sal.res.2, !is.na(Life.History)), aes(x = reorder(Life.History, pglsres.cdmass, fun=mean), y=pglsres.cdmass, fill=Life.History, shape=Life.History)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA, alpha=0.5) +
  stat_summary(fun.y = mean, colour="black", geom="point", size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species, shape = Life.History), size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Relative cornea size") +
  xlab("Life History") +
  theme(legend.position = "none") +
# use our custom colors for habitat
  scale_fill_manual(values = col_lh, name = "Life History") +
# use our custom shapes for points
  scale_shape_manual(values = shape_lh, name = "Life History") 

#print plot
lifehistory.rel

#look at the plot with the plotly package - you can hover over data points to see what each species is, great way to check outliers!
library(plotly)
ggplotly(lifehistory.rel)

```

##Life History PGLS
```{r}
# Run Life History pgls--------

#PGLS cornea diameter vs. cube root of mass + life history
pgls_lh <- pgls(log10(cor_av) ~ log10(rootmass_av) + Life.History,
               data = sal.comp, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

#diagnostic plots
par(mar = c(4,4,2,2))
par(mfrow = c(2, 2))
plot(pgls_lh)

par(mfrow = c(1, 1))

#Likelihood plot for Pagel's lambda. Solid red line indicates estimate for lambda and broken red lines indcaite the 95% confidence interval
lambda.lh <- pgls.profile(pgls_lh, "lambda")
plot(lambda.lh)

```

```{r}
#main effects
anova(pgls_lh)

#print model output 
summary(pgls_lh)
```

#Larval Habitat

#Larval Habitat Plot

```{r}
#see what the levels of development are named
levels(as.factor(sal.res.2$Larval.Habitat))

#define a colorblind-friendly vector of colors for adult habitat
col_larv <- c("lentic water" = "#CC3366",
             "lotic water" = "#CC0000",
             "no larva" = "#FF9933",
             "both"="#CC9900")

#now you can see that each habitat is assigned a hex color when you look at the vector
col_larv

#make a vector of point shapes for adult habitat
shape_larv <- c("lentic water" = 25,
             "lotic water" = 24,
             "no larva" = 23,
             "both" = 22)

#see that each state has a shape now
shape_larv
```

```{r}
#boxplot of pgls residuals vs. development (relative eye size)
larvalhabitat.rel <- ggplot(data = filter(sal.res.2, !is.na(Larval.Habitat)), aes(x = reorder(Larval.Habitat, pglsres.cdmass, fun=mean), y=pglsres.cdmass, fill=Larval.Habitat, shape=Larval.Habitat)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA, alpha=0.5) +
  stat_summary(fun.y = mean, colour="black", geom="point", size=3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species, shape = Larval.Habitat), size = 1.3, alpha = 0.3, position = position_jitter(0.15)) +
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("Relative cornea size") +
  xlab("Larval Habitat") +
  theme(legend.position = "none") +
# use our custom colors for habitat
  scale_fill_manual(values = col_larv, name = "Larval Habitat") +
# use our custom shapes for points
  scale_shape_manual(values = shape_larv, name = "Larval Habitat") 

#print plot
larvalhabitat.rel

#look at the plot with the plotly package - you can hover over data points to see what each species is, great way to check outliers!
library(plotly)
ggplotly(larvalhabitat.rel)

```

#Larval Habitat PGLS

```{r}
# Run Larval Habitat pgls--------

#PGLS cornea diameter vs. cube root of mass + larval habitat


pgls_larv <- pgls(log10(cor_av) ~ log10(rootmass_av) + Larval.Habitat,
               data = sal.comp, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)


#diagnostic plots
par(mar = c(4,4,2,2))
par(mfrow = c(2, 2))
plot(pgls_larv)

par(mfrow = c(1, 1))

#Likelihood plot for Pagel's lambda. Solid red line indicates estimate for lambda and broken red lines indcaite the 95% confidence interval

lambda.larv <- pgls.profile(pgls_larv, "lambda")
plot(lambda.larv)

```

```{r}
#main effects
anova(pgls_larv)

#print model output 
summary(pgls_larv)
``` 

```{r}
sal.phy %>%
   group_by(hab) %>%
   summarise(average = mean(cor_av),
             count = n())
```
