==> Rmd Files/8-pupil_analysis.Rmd <==
---
title: "Pupil Analysis"
author: "Nicole Bednarik & Katie Thomas"
date: 27 March 2023
output:
  html_document:
    theme: flatly
    toc: TRUE
    toc_float: TRUE
    code_fold: show
---

```{r setup, include = FALSE}
# Load package libraries
library(plyr)
library(geiger)
library(caper)
library(plotly)
library(kableExtra)
library(gridExtra)
library(ggimage)
library(grid)
library(gtable)
library(cowplot)
library(knitr)
library(tidyverse)
# Markdown settings
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, fig.height = 4.5, fig.width=8)
```

# Data

First we import our tidied salamander morphology & trait binary dataset with the residuals of a cornea ~ cube root of mass PGLS as a column.

```{r load-data}
# Import tidied salamander eye size and trait data
pglsres.cdmass.all <- data.frame(read.csv("../Data/Tidy/pglsres.cdmass.all.csv", header=TRUE, na.strings=c("", "NA", " ","<1")))
pupils <- data.frame(read.csv("../Data/Raw/pupils.csv", header=TRUE, na.strings=c("", "NA", " ","<1")))
caudatatree <- read.nexus("../Data/Tidy/caudata-tree.nex")

# Quick look at data structure
str(pglsres.cdmass.all)

# Reorder factor levels
pglsres.cdmass.all$Activity_Period <- factor(pglsres.cdmass.all$Activity_Period, levels = c("nocturnal", "diurnal", "both"))
pglsres.cdmass.all$Larval_Habitat <- factor(pglsres.cdmass.all$Larval_Habitat, levels = c("lentic water", "lotic water", "both", "no larva"))

# Merge datasets and pupil factor levels
pupils_all <- left_join(pglsres.cdmass.all, pupils, by = "Genus_Species")
pupils_all$Pupils[pupils_all$Pupils %in% c("horizontal_oval", "horizontal_slit", "horizontal_upside_down_triangle")] <- "horizontal"

#Remove NA values
pupils_all <- pupils_all[complete.cases(pupils_all$Pupils), ]

# use caper function to combine phylogeny and data into one object (this function also matches species names in tree and dataset)
sal.comp <- comparative.data(phy = caudatatree, data = pupils_all, 
                            names.col = Genus_Species, vcv = TRUE, 
                            na.omit = FALSE, warn.dropped = TRUE)

#check for dropped tips or dropped species
sal.comp$dropped$tips #phylogeny
```

# Defining colors for pupil shape

```{r}
#see what the levels of pupil shape are named
levels(as.factor(pupils_all$Pupils))

#define a colorblind-friendly vector of colors for pupil shape
col_pupil <- c("horizontal" = "#332288",
             "symmetrical_circle" = "#AA4499")

#now you can see that each habitat is assigned a hex color when you look at the vector
col_pupil

#make a vector of point shapes for pupil shape
shape_pupil <- c("horizontal" = 25,
             "symmetrical_circle" = 24)

#see that each state has a shape now
shape_pupil
```

# Testing for differences between cornea size and pupil shape

```{r}
# Boxplot of cornea diameter vs pupil shape
cor_pup <- ggplot(data = pupils_all, aes(x = reorder(Pupils, cor_av, FUN = mean), y = cor_av)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA, alpha = 0.5) +
  stat_summary(fun.y = mean, color = "black", geom = "point", 
               shape = 18, size = 3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species, shape = Pupils, color = Pupils), 
              shape = 19, size = 2, alpha = 0.5, position = position_jitter(0.15)) +
  theme(text = element_text(size = 14), panel.background = element_blank(), 
        axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  scale_color_manual(values = col_pupil) +  # Use the color vector for manual color coloring based on pupil shape
  ylab("Cornea diameter (mm)") +
  xlab("Pupil shape") +
  theme(legend.position = "none")
ggplotly(cor_pup)

# Boxplot of cornea investment (by mass) vs pupil shape
inv_pup <- ggplot(data = pupils_all, aes(x = reorder(Pupils, pglsres.cdmass, FUN = mean), y = pglsres.cdmass)) +
  geom_boxplot(notch = FALSE, outlier.shape = NA, alpha = 0.5) +
  stat_summary(fun.y = mean, color = "black", geom = "point", 
               shape = 18, size = 3, show_guide = FALSE) +
  geom_jitter(aes(text = Genus_Species, color = Pupils), 
              shape = 19, size = 2, alpha = 0.5, position = position_jitter(0.15)) +
  theme(text = element_text(size = 14), panel.background = element_blank(), 
        axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  scale_color_manual(values = col_pupil) +  # Use the color vector for manual color coloring based on Larval Habitat
  ylab("Cornea investment (mass)") +
  xlab("Pupil shape") +
  theme(legend.position = "none")
ggplotly(inv_pup)

# Export plot to pdf - you can change the export dimensions by changing the height and width numbers to make it fit well on your poster!
ggsave("../Plots/pupils.abs.boxplot.pdf", plot = cor_pup, device = "pdf")
ggsave("../Plots/pupils.inv.boxplot.pdf", plot = inv_pup, device = "pdf")


```

```{r}
# Convert 'Pupils' from character to factor with specified levels
pupils_all$Pupils <- factor(pupils_all$Pupils, levels = c("horizontal", "symmetrical_circle"))
help <- lm(data = pupils_all, pglsres.cdmass ~ Pupils)
summary(help)
anova(help)
# Run pupil pgls--------

pgls_pupil1 <- pgls(cor_av ~ Pupils,
               data = sal.comp, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

#main effects
anova(pgls_pupil1)

#print model output 
summary(pgls_pupil1)

#PGLS cornea diameter vs. cube root of mass + pupils
pgls_pupils <- pgls(log10(cor_av) ~ log10(rootmass_av) + Pupils,
               data = sal.comp, 
               lambda = "ML", #uses Maximum Likelihood estimate of lambda
               param.CI = 0.95)

#diagnostic plots
par(mar = c(4,4,2,2))
par(mfrow = c(2, 2))
plot(pgls_pupils)
par(mfrow = c(1, 1))

#Likelihood plot for Pagel's lambda. Solid red line indicates estimate for lambda and broken red lines indcaite the 95% confidence interval
lambda.pupils <- pgls.profile(pgls_pupils, "lambda")
plot(lambda.pupils)

#main effects
anova(pgls_pupils)

#print model output 
summary(pgls_pupils)

```

# Testing for differences in pupil shape across ecological states

## Adult Habitat by pupil shape

```{r}
#Adult Habitat by pupil shape
pup.hab <- ggplot(data = filter(pupils_all, !is.na(Pupils)), aes(x=Adult_Habitat, fill=Pupils)) +
  geom_bar(position="dodge") + 
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("count") +
  xlab("Adult Habitat") +
  theme(legend.position = "right") +
# use our custom colors for pupils
  scale_fill_manual(values = col_pupil, name = "Pupil shape") 

#print plot
pup.hab

#look at the plot with the plotly package - you can hover over data points to see what each species is, great way to check outliers!
library(plotly)
ggplotly(pup.hab)

```

```{r}
#Chi-square test
pupils.chi <- chisq.test(table(pupils_all$Pupils, pupils_all$Adult_Habitat))
print(pupils.chi)
#Check expected frequencies (should be >5)
pupils.chi$expected

```

## Larval Habitat by pupil shape

```{r}
#Larval Habitat by pupil shape
pup.larv <- ggplot(data = filter(pupils_all, !is.na(Pupils)), aes(x=Larval_Habitat, fill=Pupils)) +
  geom_bar(position="dodge") + 
  theme(text = element_text(size=14), panel.background = element_blank(), axis.line = element_line(colour = "black")) +
  theme(axis.text = element_text(angle = 0, size = 9)) +
  ylab("count") +
  xlab("Larval Habitat") +
  theme(legend.position = "right") +
# use our custom colors for pupils
  scale_fill_manual(values = col_pupil, name = "Pupil shape") 

#print plot
pup.larv

#look at the plot with the plotly package - you can hover over data points to see what each species is, great way to check outliers!
library(plotly)
ggplotly(pup.larv)

```

```{r}
#Chi-square test
pupilslarv.chi <- chisq.test(table(pupils_all$Pupils, pupils_all$Larval_Habitat))
print(pupilslarv.chi)
#Check expected frequencies (should be >5)
pupilslarv.chi$expected

```
